+++ elmyra.ip.access.epo/CHANGES.rst
@@ -8,6 +8,7 @@ development
 0.10.1
 ======
 - ui: make "modified" attribute humanly readable in project chooser widget
+- ui, storage: automatically update the "project.modified" attribute when manipulating the basket
 
 0.10.0
 ======